+++ packages/evm/lib/index.ts
@@ -1,10 +1,15 @@
 export { EvmChainProvider } from './chain/EvmChainProvider';
 export { EvmMulticallProvider } from './chain/EvmMulticallProvider';
 export { OptimismChainProvider } from './chain/OptimismChainProvider';
+export { EIP1559FeeApiProvider } from './fee/EIP1559FeeApiProvider';
 export { EIP1559FeeProvider } from './fee/EIP1559FeeProvider';
 export { RpcFeeProvider } from './fee/RpcFeeProvider';
+export { EnsProvider } from './naming/EnsProvider';
 export { EvmNetworks } from './networks';
+export { CovalentNftProvider } from './nft/CovalentNftProvider';
 export { EvmNftProvider } from './nft/EvmNftProvider';
+export { MoralisNftProvider } from './nft/MoralisNftProvider';
+export { OpenSeaNftProvider } from './nft/OpenSeaNftProvider';
 export { EvmBaseSwapProvider } from './swap/EvmBaseSwapProvider';
 export { EvmSwapProvider } from './swap/EvmSwapProvider';
 export * as Typechain from './typechain';