+++ packages/evm/lib/types.ts
@@ -1,4 +1,5 @@
 import { AddressType, FeeType, TransactionRequest } from '@chainify/types';
+import { Fragment, JsonFragment } from '@ethersproject/abi';
 import { BlockWithTransactions as EthersBlockWithTransactions } from '@ethersproject/abstract-provider';
 import { PopulatedTransaction as EthersPopulatedTransaction } from '@ethersproject/contracts';
 import { Block as EthersBlock, TransactionResponse as EthersTransactionResponse } from '@ethersproject/providers';
@@ -45,4 +46,20 @@ export { EthersTransactionResponse, EthersBlock, EthersBlockWithTransactions, Et
 export enum NftTypes {
     ERC721 = 'ERC721',
     ERC1155 = 'ERC1155',
+}
+
+export type NftProviderConfig = {
+    url: string;
+    apiKey: string;
+};
+
+export type MoralisConfig = NftProviderConfig & {
+    appId: string;
+};
+
+export interface MulticallData {
+    target: string;
+    abi: ReadonlyArray<Fragment | JsonFragment | string>;
+    name: string;
+    params: ReadonlyArray<Fragment | JsonFragment | string>;
 }