+++ test/integration/clients/evm/clients.ts
@@ -10,6 +10,7 @@ import { EIP1559MockFeeProvider } from './mock/EIP1559MockFeeProvider';
 function getEvmClient(network: Network) {
     const config = EVMConfig(network);
     const provider = new providers.StaticJsonRpcProvider(network.rpcUrl);
+    // using mainnet gas fees
     const feeProvider = new EIP1559MockFeeProvider(provider);
     const chainProvider = new EVM.EvmChainProvider(network, provider, feeProvider);
     // we don't have multicall on the common address on Ganache